refactror: ユーティリティのテストを分割
